Выносит настроку фильтра в отдельный файл
