+++ js/main.js
@@ -2,3 +2,4 @@ import './util.js';
 import './data.js';
 import './card.js';
 import './form.js';
+import './filter.js';
